Refractering of code
